idp/web/requirements.txt
flask
pyyaml
redis
python-jose
ldap3
